# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Mazayada API Documentation'
  description: 'Mazayada mobile API (v1). Token-authenticated JSON API for the Flutter client covering auctions and the citizen dashboard. All responses use a unified envelope ({data, message, meta} on success; {message, errors} on error). Money is returned in dinars; send Accept-Language (ar|fr|en) to localize responses.'
  version: 1.0.0
servers:
  -
    url: 'https://mazayada.findosystem.com'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'Obtain an <b>access token</b> from <code>POST /api/v1/auth/login</code> or <code>POST /api/v1/auth/verify-otp</code> and send it as <code>Authorization: Bearer &lt;access_token&gt;</code>. Access tokens are short-lived — use the <b>refresh token</b> at <code>POST /api/v1/auth/refresh</code> (send the refresh token as the bearer) to obtain a new pair.'
security:
  -
    default: []
paths:
...
  /api/v1/auth/password/request:
    post:
      summary: 'Request password reset'
      operationId: requestPasswordReset
      description: "Step 1 — email a reset code if the NIN+email pair matches an account. The\nresponse is always neutral (does not reveal whether the account exists)."
      parameters:
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nin:
                  type: string
                  description: ''
                  example: consequatur
                email:
                  type: string
                  description: 'يجب أن يكون value بريداً إلكترونياً صحيحاً.'
                  example: carolyne.luettgen@example.org
              required:
                - nin
                - email
      security: []
  /api/v1/auth/password/verify:
    post:
      summary: 'Verify password reset'
      operationId: verifyPasswordReset
      description: "Step 2 — verify the reset code and set a new password. All existing tokens\nand sessions are revoked."
      parameters:
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nin:
                  type: string
                  description: ''
                  example: consequatur
                email:
                  type: string
                  description: 'يجب أن يكون value بريداً إلكترونياً صحيحاً.'
                  example: carolyne.luettgen@example.org
                otp:
                  type: string
                  description: 'يجب أن يكون عدد أحرف value 6 حرفاً.'
                  example: fuudtd
                password:
                  type: string
                  description: ''
                  example: 'O[2UZ5ij-e/dl4m{o,'
              required:
                - nin
                - email
                - otp
                - password
      security: []
  /api/v1/auth/recover/reveal:
    post:
      summary: 'Reveal secret question'
      operationId: revealSecretQuestion
      description: "Step 1 of account recovery — return the account's security question."
      parameters:
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nin:
                  type: string
                  description: ''
                  example: consequatur
                email:
                  type: string
                  description: 'يجب أن يكون value بريداً إلكترونياً صحيحاً.'
                  example: carolyne.luettgen@example.org
              required:
                - nin
                - email
      security: []
...
  /api/v1/auth/verify-otp:
    post:
      summary: 'Verify OTP'
      operationId: verifyOTP
      description: 'Verify the registration code and receive an access + refresh token pair.'
      parameters:
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: 'Must match an existing stored value.'
                  example: consequatur
                otp:
                  type: string
                  description: 'يجب أن يكون عدد أحرف value 6 حرفاً.'
                  example: mqeopf
                device_name:
                  type: string
                  description: 'يجب ألا يزيد عدد أحرف value عن 80 حرفاً.'
                  example: uudtdsufvyvddqamniihf
                  nullable: true
              required:
                - user_id
                - otp
      security: []
  /api/v1/auth/resend-otp:
    post:
      summary: 'Resend OTP'
      operationId: resendOTP
      description: 'Re-issue a fresh registration code, throttled by a 60-second cooldown.'
      parameters:
        -
          in: header
          name: Accept-Language
          description: ''
          example: ar
          schema:
            type: string
      responses: {  }
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  description: 'Must match an existing stored value.'
                  example: consequatur
              required:
                - user_id
      security: []
...